Massive improvement of general graphics
I've completely changed the graphics of the topics layout.
Now we have the topic on the left, and the proposals on the right.
Proposals are now scrollable.
Minor changes on the view of topic and proposals details.
In home page, the groups are now scrollable.
I've added the Home option in the navbar.
I moved the 'plus' buttons.
We have to discuss about all these changes.